All error tags in the same line now become more transparent when hovering on them
